pyproject.toml
[tool.poetry]
name = "cw-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.3"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"
requests = "^2.32.3"
stripe = "^9.10.0"
celery = "^5.4.0"
redis = "^5.0.5"
django-celery-beat = "^2.6.0"
eventlet = "^0.36.1"
django-filter = "^24.2"
django-cors-headers = "^4.3.1"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
